chore(tools): include changelog in release-it tag annotation
Also, the command is run in WSL as windows' cmd.exe gets confused with
the new lines in the command due to the changelog.
Note: The `||` at the end of the `changelogCommand` is to force the
command to exit with 0. Grep exits with 1 if no lines match, and that
would cause the whole changelog command to fail.